migrate vips.rb from science to core
vips.rb is currently in homebrew-science.

We've recently done a php binding, but homebrew-php are reluctant to
add a dependency on a non-core formula. They've suggested I should
ask for vips.rb to be moved into homebrew-core:

Homebrew/homebrew-php#3757

This has previously been discussed in this PR:

https://github.com/Homebrew/homebrew-core/pull/11601

This delete PR is linked to an add PR in homebrew-core:

https://github.com/Homebrew/homebrew-core/pull/14300

There is no version change or change to any deps.